mdeliver: cast printing of time_t

+++ mdeliver.c
@@ -79,7 +79,8 @@ tryagain:
 		gettimeofday(&tv, 0);
 
 		snprintf(id, sizeof id, "%ld.M%06ldP%ldQ%ld.%s",
-			 tv.tv_sec, tv.tv_usec, (long)getpid(), delivery, host);
+		    (long)tv.tv_sec, (long)tv.tv_usec, (long)getpid(),
+		    delivery, host);
 		
 		snprintf(tmp, sizeof tmp, "%s/tmp/%s", targetdir, id);
